refactor unit test code
